file_organization_backup/backup_20250616_121023/file_organization_backup/backup_20250616_121023/file_organization_backup/backup_20250616_121023/code_backups/backup_20250616_105453/requirements.txt
# 핵심 라이브러리만 유지
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
websocket-client>=1.5.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
schedule>=1.2.0
asyncio>=3.4.3
yfinance>=0.2.0
aiohttp>=3.8.0

# 개발 도구
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
bandit>=1.7.0
safety>=2.3.0 